pyproject.toml
[project]
name = "camar"
version = "0.2.0"
description = "CAMAR: A high-performance multi-agent reinforcement learning environment for continuous multi-agent pathfinding"
readme = "README.md"
authors = [
    { name = "Square596", email = "artem211201@gmail.com" }
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license = "MIT"
license-files = ["LICENSE"]

dependencies = [
    "flax>=0.9.0",
    "jax>=0.4.31",
]

[project.optional-dependencies]
cuda12 = [
    "jax[cuda12]>=0.4.31",
]
labmaze = [
    "labmaze>=1.0.6",
]
movingai = [
    "numpy>=1.24.0",
    "opencv-python>=4.5.4.60",
    "requests>=2.20.0",
]
matplotlib = [
    "matplotlib>=3.10.1",
]
torchrl = [
    "torchrl>=0.7.0",
]
baselines = [
    "matplotlib>=3.10.1",
    "pandas>=1.4.4",
    "torchrl>=0.7.0",
    "tqdm>=4.67.1",
    "wandb>=0.19.9",
]
benchmarl = [
    "av<14.0.0",
    "torchrl>=0.7.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "hydra-core>=1.3.2",
    "wandb>=0.19.9",
    "id-marl-eval>=0.0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "pytest>=8.0.0",
    "pyyaml>=6.0.2",
    "ruff>=0.11.13",
]

[tool.ruff]
line-length = 110